Refector InputView : change struct name
